+++ src/main/java/roomescape/resolver/JWTBearerTokenSubjectResolver.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import roomescape.annotation.JWTBearerTokenSubject;
 import roomescape.exception.AuthorizationException;
 import roomescape.service.JWTProvider;
